renamed files, cleanup
